refactor: starts modularization and unit testing
