+++ .gitignore
@@ -1,3 +1,5 @@
-
+__pycache__
+*.pyc
 dist/*
 Tests/*
+tags
