+++ requirements-dev.txt
@@ -1,0 +1,1 @@
+pytest==6.1.2
