+++ requirements.txt
@@ -1,0 +1,26 @@
+cachetools==4.2.2
+certifi==2021.5.30
+cffi==1.14.5
+chardet==4.0.0
+DateTime==4.3
+google-api-core==1.23.0
+google-auth==1.23.0
+google-cloud-core==1.4.4
+google-cloud-storage==1.32.0
+google-crc32c==1.0.0
+google-resumable-media==1.1.0
+googleapis-common-protos==1.52.0
+idna==2.10
+protobuf==3.17.2
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+pycparser==2.20
+PyMySQL==0.10.1
+pytz==2021.1
+requests==2.25.1
+rsa==4.7.2
+selenium==3.141.0
+six==1.16.0
+slacker==0.14.0
+urllib3==1.26.5
+zope.interface==5.4.0
