+++ tests/unit/test_airtable.py
@@ -1,0 +1,9 @@
+import unittest
+
+from NathanJamesToolbox import Airtable, airtableToolbox
+
+
+class TestAirtable(unittest.TestCase):
+
+    def test_legacy_interface(self):
+        self.assertIs(airtableToolbox, Airtable)
